"register"

+++ js/register-page.js
@@ -3,6 +3,7 @@ roleNamesDrop = document.getElementById("role-names");
 for (let i = 0; i < roleNames.length; ++i) {
     roleNamesDrop[roleNamesDrop.length] = new Option(roleNames[i], roleNames[i]);
 }
+
 document.getElementById("registerButton").addEventListener("click", register); 
 
 
@@ -19,8 +20,18 @@ function register(){
         "password" : password,
         "role" : roleValue
     };
-    existingUsers.push(newUser);
-    localStorage.setItem("users", JSON.stringify(existingUsers));
-    console.log(localStorage.getItem("users"));
-    //window.location.href="../html/home.html";
+
+    if(!existingUsers.some(user => user["username"] === username)){
+        existingUsers.push(newUser);
+        console.log("account created");
+        localStorage.setItem("users", JSON.stringify(existingUsers));
+        console.log(localStorage.getItem("users"));
+        alert("Successful! You can now log in.");
+        window.location.href="../html/home.html";
+    }else{
+        alert("Username exists in database.");
+        console.log(localStorage.getItem("users"));
+    }
+
+   
 }
